Rename entities to decorators
In preparation of https://github.com/springload/draftail/issues/49
