+++ wagtaildraftail/client/decorators/Document.js
