+++ wagtaildraftail/client/decorators/index.js
