+++ wagtaildraftail/client/decorators/index.test.js
