+++ wagtaildraftail/client/wagtaildraftail.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Entity } from 'draft-js';
-import DraftailEditor, { ENTITY_TYPE } from 'draftail';
+import DraftailEditor from 'draftail';
 
 import 'draftail/dist/draftail.css';
 import './wagtaildraftail.css';
 
 import sources from './sources';
-import entities from './entities';
+import decorators from './decorators';
 
 // TODO: Use the one from draftail once implemented https://github.com/springload/draftail/issues/48
 const getEntityStrategy = (entityType) => {
@@ -34,9 +34,10 @@ const initDraftailEditor = (fieldName, options = {}) => {
   if (options.entityTypes) {
     // eslint-disable-next-line no-param-reassign
     options.entityTypes = options.entityTypes.map(entity => Object.assign(entity, {
+      // TODO: Rename keys accordingly once changed in draftail https://github.com/springload/draftail/issues/49
       control: sources[entity.source],
       strategy: getEntityStrategy(entity.type),
-      component: entities[entity.decorator],
+      component: decorators[entity.decorator],
     }));
   }
